upload prob3

+++ ritidet/paramita/lab2/DataTypes.java
@@ -23,8 +23,8 @@ public class DataTypes {
         boolean valueOfBoolean = true ;
         int myID_in_Octal = 0111 ;
         int myID_in_Hexadecimal = 0x49 ;
-        long lastTwoDigits_Of_ID = 73 ;
-        float lastTwoDigitsID_point_FirstTwoDigitsID = (float) 73.65 ;
+        long lastTwoDigits_Of_ID = 73L ;
+        float lastTwoDigitsID_point_FirstTwoDigitsID = 73.65F ;
         Double lastTwoDigitsID_point_LastTwoDigitsCurrentYear = 73.22 ;
         System.out.println("My name is " + my_name);
         System.out.println("My student ID was " + my_studentID);
